add Dashboard page with participants table
