+++ views/tables/ConfigurationsTable/index.jsx
@@ -21,6 +21,7 @@ const ConfigurationsTable = (props) => {
     onEdit,
     onShare,
     user,
+    maxRows,
   } = props
   const headers = [
     { dataProperty: 'name', label: 'Configuration Name', sortable: false },
@@ -107,6 +108,7 @@ const ConfigurationsTable = (props) => {
       cellRenderer={cellRenderer}
       data={configurations}
       headers={headers}
+      maxRows={maxRows}
     />
   )
 }
